engine/requirements.txt
# Prediction Market Decision Engine
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0

# Price Data
yfinance>=0.2.0

# Polymarket
py-clob-client>=0.1.0

# News & Sentiment
feedparser>=6.0.0
vaderSentiment>=3.3.2

# Streamlit Dashboard
streamlit>=1.28.0
plotly>=5.15.0
plotly>=5.15.0
pandas>=2.0.0
